+++ askbot/startup_procedures.py
@@ -251,7 +251,7 @@ def test_template_loader():
         errors.append(
             '"%s" must be the first element of TEMPLATE_LOADERS' % current_loader
         )
-        
+
     print_errors(errors)
 
 def test_celery():
@@ -683,7 +683,6 @@ def test_tinymce():
     required_attrs = (
         'TINYMCE_COMPRESSOR',
         'TINYMCE_JS_ROOT',
-        'TINYMCE_URL',
         'TINYMCE_DEFAULT_CONFIG'
     )
 
@@ -738,16 +737,6 @@ def test_tinymce():
             error_tpl += '\nNote: we have moved files from "common" into "default"'
         errors.append(error_tpl % relative_js_path)
 
-    #check url setting
-    url = getattr(django_settings, 'TINYMCE_URL', '')
-    expected_url = django_settings.STATIC_URL + relative_js_path
-    old_expected_url = django_settings.STATIC_URL + old_relative_js_path
-    if urls_equal(url, expected_url) is False:
-        error_tpl = "add line: TINYMCE_URL = STATIC_URL + '%s'"
-        if urls_equal(url, old_expected_url):
-            error_tpl += '\nNote: we have moved files from "common" into "default"'
-        errors.append(error_tpl % relative_js_path)
-
     if errors:
         header = 'Please add the tynymce editor configuration ' + \
             'to your settings.py file.'
@@ -795,7 +784,7 @@ def test_template_context_processors():
         required_processors.append(new_auth_processor)
         if old_auth_processor in django_settings.TEMPLATE_CONTEXT_PROCESSORS:
             invalid_processors.append(old_auth_processor)
-            
+
     missing_processors = list()
     for processor in required_processors:
         if processor not in django_settings.TEMPLATE_CONTEXT_PROCESSORS:
@@ -864,7 +853,7 @@ def test_group_messaging():
             errors.append(
                 "make setting 'GROUP_MESSAGING to be exactly:\n" + settings_sample
             )
-            
+
         url_params = settings.get('BASE_URL_PARAMS', None)
     else:
         errors.append('add this to your settings.py:\n' + settings_sample)
@@ -893,7 +882,7 @@ def test_multilingual():
         errors.append('ASKBOT_MULTILINGUAL=True works only with django >= 1.4')
 
     if is_multilang:
-        middleware = 'django.middleware.locale.LocaleMiddleware' 
+        middleware = 'django.middleware.locale.LocaleMiddleware'
         if middleware not in django_settings.MIDDLEWARE_CLASSES:
             errors.append(
                 "add 'django.middleware.locale.LocaleMiddleware' to your MIDDLEWARE_CLASSES "
@@ -909,6 +898,10 @@ def test_multilingual():
 
     print_errors(errors)
 
+def test_messages_framework():
+    if not 'django.contrib.messages' in django_settings.INSTALLED_APPS:
+        errors = ('Add to the INSTALLED_APPS section of your settings.py:\n "django.contrib.messages"', )
+        print_errors(errors)
 
 def run_startup_tests():
     """function that runs
@@ -932,6 +925,7 @@ def run_startup_tests():
     test_new_skins()
     test_media_url()
     #test_postgres()
+    test_messages_framework()
     test_middleware()
     test_multilingual()
     #test_csrf_cookie_domain()