+++ askbot/tests/search_state_tests.py
@@ -64,7 +64,7 @@ class SearchStateTests(AskbotTestCase):
 
     def test_edge_cases_1(self):
         ss = SearchState(
-            scope='favorite', # this is not a valid choice for non-logger users
+            scope='followed', # this is not a valid choice for non-logger users
             sort='age-desc',
             query=' alfa',
             tags='miki, mini',
@@ -83,7 +83,7 @@ class SearchStateTests(AskbotTestCase):
         )
 
         ss = SearchState(
-            scope='favorite',
+            scope='followed',
             sort='age-desc',
             query=' alfa',
             tags='miki, mini',
@@ -93,11 +93,11 @@ class SearchStateTests(AskbotTestCase):
             user_logged_in=True
         )
         self.assertEqual(
-            'scope:favorite/sort:age-desc/query:alfa/tags:miki,mini/author:12/page:2/',
+            'scope:followed/sort:age-desc/query:alfa/tags:miki,mini/author:12/page:2/',
             ss.query_string()
         )
         self.assertEqual(
-            'scope:favorite/sort:age-desc/query:alfa/tags:miki,mini/author:12/page:2/',
+            'scope:followed/sort:age-desc/query:alfa/tags:miki,mini/author:12/page:2/',
             ss.deepcopy().query_string()
         )