+++ askbot/doc/source/changelog.rst
@@ -4,6 +4,7 @@ Changes in Askbot
 Development version
 -------------------
 * Some support for the media compression (Tyler Mandry)
+* Allowed to enable and disable question scopes on the main page
 * Added full text support for some languages with Postgresql:
   Danish, Dutch, English, Finnish, French, German, Hungarian,
   Italian, Japanese (requires package textsearch_ja), Norwegian,